feat: testhelper.ImportCsv function
